requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
anytree=2.8.0=pypi_0
apex=0.1=pypi_0
attrs=20.3.0=pyhd3deb0d_0
backcall=0.2.0=py_0
blas=1.0=mkl
blis=0.4.1=pypi_0
boto3=1.16.23=pypi_0
botocore=1.19.23=pypi_0
brotlipy=0.7.0=py38h8df0ef7_1001
ca-certificates=2020.12.8=h06a4308_0
catalogue=1.0.0=py38h578d9bd_2
certifi=2020.11.8=pypi_0
cffi=1.14.3=py38h261ae71_2
chardet=3.0.4=py38h924ce5b_1008
click=7.1.2=py_0
cloudpickle=1.6.0=py_0
cryptography=3.2.1=py38h7699a38_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py38_0
cymem=2.0.4=py38hadf7658_1
cython=0.29.21=pypi_0
cython-blis=0.4.1=py38h1e0a361_1
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2.30.0=py_0
decorator=4.4.2=py_0
en-core-web-sm=2.3.1=pypi_0
freetype=2.10.4=h5ab3b9f_0
h5py=3.1.0=pypi_0
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=py_0
importlib-metadata=3.0.0=pyhd8ed1ab_0
importlib_metadata=3.0.0=hd8ed1ab_0
intel-openmp=2020.2=254
ipython=7.19.0=py38hb070fc8_0
ipython_genutils=0.2.0=py38_0
jedi=0.17.2=py38_0
jmespath=0.10.0=pypi_0
joblib=0.17.0=py_0
jpeg=9b=h024ee3a_2
jsonschema=3.2.0=py_2
kiwisolver=1.3.0=py38h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
lz4-c=1.9.2=heb0550a_3
matplotlib-base=3.3.2=py38h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.2.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
murmurhash=1.0.4=pypi_0
ncurses=6.2=he6710b0_1
networkx=2.5=py_0
ninja=1.10.0.post2=pypi_0
nltk=3.5=py_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
olefile=0.46=py_0
opencv-python=4.4.0.46=pypi_0
openssl=1.1.1i=h27cfd23_0
pandas=1.1.3=py38he6710b0_0
parso=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=py38_1000
pillow=8.0.1=py38he98fc37_0
pip=20.2.4=py38h06a4308_0
plac=1.1.3=pypi_0
preshed=3.0.4=py38hadf7658_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.6.0=pyhd3eb1b0_2
pycocotools=2.0=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pygments=2.7.2=pyhd3eb1b0_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyrsistent=0.17.3=py38h25fe258_1
pysocks=1.7.1=py38h924ce5b_2
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.7.0=py3.8_cuda10.2.89_cudnn7.6.5_0
pytz=2020.4=pyhd3eb1b0_0
pywavelets=1.1.1=py38h7b6447c_2
pyyaml=5.3.1=py38h7b6447c_1
readline=8.0=h7b6447c_0
regex=2020.10.15=py38h7b6447c_0
requests=2.25.0=pyhd3deb0d_0
s3transfer=0.3.3=pypi_0
scikit-image=0.17.2=py38hdf5156a_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.5.2=py38h0b6359f_0
setuptools=50.3.1=py38h06a4308_1
six=1.15.0=py38h06a4308_0
spacy=2.3.2=py38hbf85e49_0
spacy-lookups-data=0.3.0=pyh9f0ad1d_0
sqlite=3.33.0=h62c20be_0
srsly=1.0.4=py38hadf7658_0
thinc=7.4.1=py38h82cb98a_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py38hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=py_0
torchvision=0.8.1=py38_cu102
tornado=6.0.4=py38h7b6447c_1
tqdm=4.51.0=pyhd3eb1b0_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.2=pypi_0
wasabi=0.8.0=pyh9f0ad1d_0
wcwidth=0.2.5=py_0
wheel=0.35.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yacs=0.1.8=pypi_0
yaml=0.2.5=h7b6447c_0
zipp=3.4.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
